requirements.txt
requests
beautifulsoup4
mysql-connector-python
prefect
python-dotenv
